packages/wabizabi/pyproject.toml
[project]
name = "wabizabi"
version = "0.1.0"
description = "Typed, streaming-first, provider-agnostic agent orchestration."
readme = "README.md"
authors = [
  { name = "aekn", email = "aekn.admal@gmail.com" },
]
requires-python = ">=3.13"
dependencies = [
  "pydantic>=2.11,<3",
]

[project.optional-dependencies]
ollama = ["ollama>=0.6,<0.7"]
otel = ["opentelemetry-api>=1.24,<2"]

[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
wabizabi = ["py.typed"]
